requirement.txt
openai==0.28
python-dotenv
Flask